+++ truffle/src/com.oracle.truffle.runtime/src/com/oracle/truffle/runtime/BytecodeOSRRootNode.java
@@ -55,13 +55,13 @@ import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.BytecodeOSRNode;
 
 final class BytecodeOSRRootNode extends BaseOSRRootNode {
-    private final int target;
+    private final long target;
     private final Object interpreterState;
     @CompilationFinal private boolean seenMaterializedFrame;
 
     private final Object entryTagsCache;
 
-    BytecodeOSRRootNode(TruffleLanguage<?> language, FrameDescriptor frameDescriptor, BytecodeOSRNode bytecodeOSRNode, int target, Object interpreterState, Object entryTagsCache) {
+    BytecodeOSRRootNode(TruffleLanguage<?> language, FrameDescriptor frameDescriptor, BytecodeOSRNode bytecodeOSRNode, long target, Object interpreterState, Object entryTagsCache) {
         super(language, frameDescriptor, bytecodeOSRNode);
         this.target = target;
         this.interpreterState = interpreterState;
@@ -101,7 +101,11 @@ final class BytecodeOSRRootNode extends BaseOSRRootNode {
             return osrNode.executeOSR(parentFrame, target, interpreterState);
         } else {
             if (usesDeprecatedFrameTransfer) { // Support for deprecated frame transfer: GR-38296
-                osrNode.copyIntoOSRFrame(frame, parentFrame, target);
+                int intTarget = (int) target;
+                if (intTarget != target) {
+                    throw CompilerDirectives.shouldNotReachHere("long target cannot be used with deprecated frame transfer");
+                }
+                osrNode.copyIntoOSRFrame(frame, parentFrame, intTarget);
             } else {
                 osrNode.copyIntoOSRFrame(frame, parentFrame, target, entryTagsCache);
             }